pyproject.toml
[tool.poetry]
name = "poe-shell"
version = "0.1.0"
description = ""
authors = ["Samson Hui <hui.kawang@yahoo.com.hk>"]
readme = "README.md"
packages = [{include = "poe_shell"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
poe-api = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
